+++ src/interface/network/net_Socket.c
@@ -19,7 +19,6 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  *******************************************************************************/
-#include <github.com/Lobaro/lobaro-coap/liblobaro_coap.h>
 #include "../../coap.h"
 
 typedef struct {
